# src/doc/swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Book List RestAPI",
    "description": "Livros e Listas para todos os gostos!"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/book/getById": {
      "get": {
        "tags": [
          "book"
        ],
        "description": "Busca livro pelo Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Id do livro a ser buscado.",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idBook": {
                      "type": "number",
                      "example": 0
                    },
                    "photo": {
                      "type": "string",
                      "example": "base64 ou null"
                    },
                    "name": {
                      "type": "string",
                      "example": "nome"
                    },
                    "publisher": {
                      "type": "string",
                      "example": "Editora"
                    },
                    "dateOfPublication": {
                      "type": "string",
                      "example": "yyyy-MM-dd"
                    },
                    "synopsis": {
                      "type": "string",
                      "example": "sinopse"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/book/searchBook": {
      "get": {
        "tags": [
          "book"
        ],
        "description": "Pesquisar livros",
        "parameters": [
          {
            "name": "term",
            "description": "Termo para buscar o livro",
            "type": "string",
            "in": "query"
          },
          {
            "name": "page",
            "description": "Página de itens.",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "itens",
            "description": "Quantidade de itens a buscar.",
            "required": true,
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "photo": {
                        "type": "string",
                        "example": "base64 ou null"
                      },
                      "name": {
                        "type": "string",
                        "example": "nome"
                      },
                      "publisher": {
                        "type": "string",
                        "example": "Editora"
                      },
                      "dateOfPublication": {
                        "type": "string",
                        "example": "yyyy-MM-dd"
                      },
                      "synopsis": {
                        "type": "string",
                        "example": "sinopse"
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/book/getLastBooks": {
      "get": {
        "tags": [
          "book"
        ],
        "description": "Busca os últimos livros criados",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "photo": {
                        "type": "string",
                        "example": "base64 ou null"
                      },
                      "name": {
                        "type": "string",
                        "example": "nome"
                      },
                      "publisher": {
                        "type": "string",
                        "example": "Editora"
                      },
                      "dateOfPublication": {
                        "type": "string",
                        "example": "yyyy-MM-dd"
                      },
                      "synopsis": {
                        "type": "string",
                        "example": "sinopse"
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/favorite/create": {
      "post": {
        "tags": [
          "favorite"
        ],
        "description": "Criar novo livro favorito do usuário.",
        "parameters": [
          {
            "name": "favorite",
            "in": "body",
            "required": true,
            "description": "Id do usuário e livro para criação de um favorito.",
            "schema": {
              "type": "object",
              "properties": {
                "idUser": {
                  "type": "number",
                  "example": 0
                },
                "idBook": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idBook": {
                      "type": "number",
                      "example": 0
                    },
                    "idUser": {
                      "type": "number",
                      "example": 0
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/favorite/getFavoritesListByIdUser": {
      "get": {
        "tags": [
          "favorite"
        ],
        "description": "Listar livros favoritos do usuário.",
        "parameters": [
          {
            "name": "idUser",
            "required": true,
            "description": "Id do usuário a ter seus favoritos listados.",
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "idUser": {
                        "type": "number",
                        "example": 0
                      },
                      "Book": {
                        "type": "object",
                        "properties": {
                          "idBook": {
                            "type": "number",
                            "example": 0
                          },
                          "photo": {
                            "type": "string",
                            "example": "base64 ou null"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome"
                          },
                          "author": {
                            "type": "string",
                            "example": "autor"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "editora"
                          },
                          "dateOfPublication": {
                            "type": "string",
                            "example": "yyyy-MM-dd"
                          },
                          "synopsis": {
                            "type": "string",
                            "example": "sinopse"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/favorite/delete": {
      "delete": {
        "tags": [
          "favorite"
        ],
        "description": "Remoção de livro como favorito do usuário.",
        "parameters": [
          {
            "name": "idUser",
            "required": true,
            "description": "Id do usuário",
            "type": "integer",
            "in": "query"
          },
          {
            "name": "idBook",
            "required": true,
            "description": "Id do livro",
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Favorito deletado com sucesso."
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}